+++ app/setup.cfg
@@ -1,12 +1,6 @@
 [flake8]
 max-line-length = 119
-exclude =
-  migrations,
-  load_data.py,
-  utils,
-  venv,
-  app/core/scripts
-
+exclude = '^(app/core/migrations/.*)$'
 max-complexity = 4
 per-files-ignore = */utils/*.*: N818
