+++ pyproject.toml
@@ -1,2 +1,12 @@
 [tool.ruff.lint]
 select = [ "PTH" ]
+[tool.black]
+exclude = '''
+/(
+  migrations|
+  venv|
+  app/core/scripts|
+  app/core/migrations|
+  app/data/migrations
+)/
+'''
